# 1. GestIC(Aries)/backend/src/config/swagger.yaml
swagger: "2.0"
info:
  description: "GestIC App API documentation"
  version: "1.0.0"
  title: "GestIC API"
host: "localhost:3333"
...
paths:
  /access/login:
    post:
      tags:
      - "User"
      summary: "Authentication and authorization of a registered user"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Parameters used in user authentication"
        required: true
        schema:
         type: "object"
         properties:
            email:
              type: "string"
            password:
              type: "string"
      responses:
        "200":
          description: "Successful login. profileTag can be ALUN or MONI or COOR or PROF."
          schema:
            allOf:
              - $ref: "#/components/schemas/User"
              - type: "object"
                properties:
                  password:
                    type: "string"
                    example: "******"
                  profileTag:
                    type: "string"
                  auth:
                    type: "boolean"
                  token:
                    example: 	"bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjA4OTI5NzE4LTU0ZGEtNDVmNy04YTVjLTQ2ZTg
                    1ZTFjNzc2YyIsInByb2ZpbGVJZCI6IjIxZjMzOTJkLWZiZGMtNGNiMC04ZTNjLTVhMTFkNmMwNjlmMSIsImlhdCI6MTYxOTA1ODc3NSwiZXhw
                    IjoxNjE5MDU5Njc1fQ.OpZp7wcmzolDnIZ12N2FVOQgrIhui6G0Ok6IEitRcUw"
        "500":
          description: "Unsuccessful login"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
  /access/register:
    post:
      tags:
      - "User"
      summary: "Register a new user"
      description: "Regis"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Information required for registration"
        required: true
        schema:
          type: "object"
          properties:
            name:
              type: "string"
            email:
              type: "string"
            password:
              type: "string"
              example: "******"
      responses:
        "200":
          description: "Successful registration"
          schema:
            allOf:
              - $ref: "#/components/schemas/User"
              - type: "object"
                properties:
                  password:
                    type: "string"
                  profileTag:
                    type: "string"
        "500":
          description: "Unsuccessful registration"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
...
  /research:
    get:
      tags:
      - "Research groups"
      summary: "Get a list of all research groups"
      produces:
      - "application/json"
      responses:
        "200":
          description: "Obtained a list of all active research groups"
          schema:
            type: "array"
            produces: "application/json"
            items:
              type: "object"
              properties:
                id:
                  type: "string"
                  format: "uuid"
                owner:
                  type: "string"
                  format: "uuid"
                name:
                  type: "string"
                description:
                  type: "string"
                activities:
                  type: "string"
        "500":
          description: "Could not obtain a list of all active research groups"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    post:
      tags:
      - "Research groups"
      summary: "Insert a research group"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: "object"
            properties:
              name:
                type: "string"
              description:
                type: "string"
              activities:
                type: "string"
      responses:
        "200":
          description: "Inserted a research group"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              activities:
                type: "string"
        "500":
          description: "Could not insert a research group"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    put:
      tags:
      - "Research groups"
      summary: " Update research group information"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          description: "Any quantity of the following parameters can be sent, but id is required."
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              activities:
                type: "string"
      responses:
        "200":
          description: "Sucessfully updated a research group"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              activities:
                type: "string"
        "500":
          description: "Could not update a research group"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
  /research/:id:
    get:
      tags:
        - "Research groups"
      summary: "Get research group information"
      produces:
      - "application/json"
      parameters:
        - in: param
          name: id
          required: true
          type: string
          format: "uuid"
      responses:
        "200":
          description: "Obtained a research group"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              activities:
                type: "string"
              members:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "string"
                      format: "uuid"
                    name:
                      type: "string"
                    email:
                      type: "string"
        "500":
          description: "Could not obtain a research group"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    delete:
      tags:
        - "Research groups"
      summary: "Delete research group"
      produces:
      - "application/json"
      parameters:
        - in: param
          name: id
          required: true
          type: string
          format: "uuid"
      responses:
        "200":
          description: "Successfully removed a research group"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              userId:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
        "500":
          description: "Could not remove a research group"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
  /research/members:
    post:
      tags:
        - "Research groups"
      summary: "Add a member to research group"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              userId:
                type: "string"
                format: "uuid"
      responses:
        "200":
          description: "Added a member"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              activities:
                type: "string"
              members:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "string"
                      format: "uuid"
                    name:
                      type: "string"
                    email:
                      type: "string"
        "500":
          description: "Could not add a member to research group"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    delete:
      tags:
        - "Research groups"
      summary: "Remove a member from research group"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              userId:
                type: "string"
                format: "uuid"
      responses:
        "200":
          description: "Removed a member"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              activities:
                type: "string"
              members:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "string"
                      format: "uuid"
                    name:
                      type: "string"
                    email:
                      type: "string"
        "500":
          description: "Could not remove a member from research group"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
...
components:
  schemas:
    User:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        profileId:
          type: "string"
          format: "uuid"
        name:
          type: "string"
        email:
          type: "string"
        status:
          type: "boolean"

      Event:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        start:
          type: "string"
          format: "date-time"
        end:
          type: "string"
          format: "date-time"
        title:
          type: "string"
        description:
          type: "string"